reimport modified micro-cosmos

+++ Teensy/agent_artemis/lib/micro-cosmos/math/crclib.cpp
@@ -1,4 +1,4 @@
-#include <math/crclib.h>
+#include "crclib.h"
 
 //! Calculate CRC-16
 /*! Calculate 16-bit CRC for the indicated type, buffer and number of bytes.
@@ -316,7 +316,7 @@ uint16_t calc_crc16(vector<uint8_t> buf, uint16_t poly, uint16_t crc, uint16_t x
 
     for (uint16_t i = 0; i < buf.size(); i++)
     {
-        uint8_t ch = buf[i];
+        uint ch = buf[i];
         for (uint16_t j = 0; j < 8; j++)
         {
             if (lsbfirst)
